landright-app/backend/requirements.txt
anthropic>=0.39.0
beautifulsoup4>=4.12.0
fastapi>=0.115.0
httpx>=0.27.0
playwright>=1.49.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.32.0
